backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.23.2
numpy==1.24.0
opencv-python==4.8.1.78
Pillow==10.0.0
ultralytics==8.0.160
transformers==4.34.0
torch==2.2.0
passporteye==0.4.3
scikit-image==0.21.0
imageio==2.31.1
